project structure refactor: moved api-related functions and modules to api_utils package
